Parsers now work directly on streams, not files
Enter and exit routines have been removed as the parsers
are not responsible for opening and closing files. The underlying
stream is now exposed but any manipulations will likely screw up
the processing.

Cleaned up some of the internals and documentation as well. The
changes to the internals of each method are not to severe, since
they already act on an opened file buffer any way.